able to run tests by url env
